+++ src/app/api/organizations/[id]/invitations/[invitationId]/route.ts
@@ -2,12 +2,15 @@ import { type NextRequest } from 'next/server';
 import InvitationService from '@/lib/services/invitation.service';
 import { updateRoleSchema } from '@/lib/schemas/role.schema';
 import { handleError } from '@/lib/utils/errors.utils';
+import { assertAdminOrOwner } from '@/lib/utils/permissions.utils';
 
 export async function PATCH(
     request: NextRequest,
     { params }: { params: Promise<{ id: string; invitationId: string }> }
 ) {
     try {
+        await assertAdminOrOwner(request.headers);
+
         const { id: organizationId, invitationId } = await params;
         const body = await request.json();
         const { role } = updateRoleSchema.parse(body);
@@ -15,8 +18,7 @@ export async function PATCH(
         const updated = await InvitationService.updatePendingInvitationRole(
             organizationId,
             invitationId,
-            role,
-            request.headers
+            role
         );
 
         return Response.json({ data: updated }, { status: 200 });
@@ -30,13 +32,11 @@ export async function DELETE(
     { params }: { params: Promise<{ id: string; invitationId: string }> }
 ) {
     try {
+        await assertAdminOrOwner(request.headers);
+
         const { id: organizationId, invitationId } = await params;
 
-        await InvitationService.deletePendingInvitation(
-            organizationId,
-            invitationId,
-            request.headers
-        );
+        await InvitationService.deletePendingInvitation(organizationId, invitationId);
 
         return new Response(null, { status: 204 });
     } catch (err) {
