+++ src/lib/components/settings/InvitationRoleCell.tsx
@@ -1,8 +1,6 @@
 'use client';
 
-import { useState } from 'react';
 import { ChevronDown } from 'lucide-react';
-import { toast } from 'sonner';
 import {
     DropdownMenu,
     DropdownMenuContent,
@@ -13,12 +11,9 @@ import {
     DropdownMenuSeparator,
     DropdownMenuTrigger,
 } from '@/lib/components/ui/Dropdown';
-import { authClient } from '@/lib/auth/auth-client';
-import { deleteOrganizationInvitation, updateInvitationRole } from '@/lib/api/organizations';
-import type { UpdateInvitationRolePayload } from '@/lib/schemas/role.schema';
+import useInvitationRoleCell from '@/lib/hooks/useInvitationRoleCell';
 import type { OrgInvitation } from '@/lib/types/invitation.types';
-import { getInvitationStatus } from '@/lib/utils/invitation.utils';
-import { getAssignableRoles, getRoleLabel, type OrgRole } from '@/lib/utils/roles.utils';
+import { getAssignableRoles, getRoleLabel } from '@/lib/utils/roles.utils';
 
 type InvitationRoleCellProps = {
     invitation: OrgInvitation;
@@ -31,62 +26,8 @@ export default function InvitationRoleCell({
     organizationId,
     onChanged,
 }: InvitationRoleCellProps) {
-    const [updating, setUpdating] = useState(false);
-    const [currentRole, setCurrentRole] = useState(invitation.role);
-
-    const handleRoleChange = async (newRole: string) => {
-        if (newRole === currentRole) return;
-        setUpdating(true);
-        const previousRole = currentRole;
-        setCurrentRole(newRole);
-        try {
-            await updateInvitationRole(organizationId, invitation.id, {
-                role: newRole as UpdateInvitationRolePayload['role'],
-            });
-            toast.success(`Invitation updated to ${getRoleLabel(newRole)}`);
-            onChanged();
-        } catch {
-            setCurrentRole(previousRole);
-            toast.error('Failed to update invitation role');
-        } finally {
-            setUpdating(false);
-        }
-    };
-
-    const handleRemoveInvitation = async () => {
-        setUpdating(true);
-        try {
-            await deleteOrganizationInvitation(organizationId, invitation.id);
-            toast.success('Invitation removed');
-            onChanged();
-        } catch {
-            toast.error('Failed to remove invitation');
-        } finally {
-            setUpdating(false);
-        }
-    };
-
-    const handleResendInvite = async () => {
-        setUpdating(true);
-        try {
-            const isExpired = getInvitationStatus(invitation) === 'invite-expired';
-            if (isExpired) {
-                await deleteOrganizationInvitation(organizationId, invitation.id);
-            }
-            await authClient.organization.inviteMember({
-                email: invitation.email,
-                role: currentRole as OrgRole,
-                organizationId,
-                resend: !isExpired,
-            });
-            toast.success('Invitation resent');
-            onChanged();
-        } catch {
-            toast.error('Failed to resend invitation');
-        } finally {
-            setUpdating(false);
-        }
-    };
+    const { updating, currentRole, handleRoleChange, handleRemoveInvitation, handleResendInvite } =
+        useInvitationRoleCell({ invitation, organizationId, onChanged });
 
     return (
         <DropdownMenu>
