+++ src/lib/components/settings/OrgLogoUploader.tsx
@@ -1,11 +1,8 @@
 'use client';
 
-import { useEffect, useState } from 'react';
 import Image from 'next/image';
 import { Loader2, Upload } from 'lucide-react';
-import { toast } from 'sonner';
-import useFileUpload from '@/lib/hooks/useFileUpload';
-import useFileClient from '@/lib/hooks/useFileClient';
+import useOrgLogoUploader from '@/lib/hooks/useOrgLogoUploader';
 
 type OrgLogoUploaderProps = {
     organization: { id: string; name: string; logo: string | null };
@@ -20,41 +17,10 @@ export default function OrgLogoUploader({
     disabled = false,
     onUpdated,
 }: OrgLogoUploaderProps) {
-    const {
-        file,
-        preview,
-        fileInputRef,
-        handleFileChange,
-        handleProfileImageClick: openFilePicker,
-    } = useFileClient();
-    const { uploadFile } = useFileUpload('organization');
-    const [isUploading, setIsUploading] = useState(false);
+    const { fileInputRef, handleFileChange, openFilePicker, isUploading, logoSrc } =
+        useOrgLogoUploader({ organization, updateLogo, onUpdated });
 
     const orgInitial = organization.name?.[0]?.toUpperCase() ?? '?';
-    const logoSrc = preview || organization.logo;
-
-    useEffect(() => {
-        if (!file) return;
-        const upload = async () => {
-            setIsUploading(true);
-            try {
-                const url = await uploadFile(file, organization.id);
-                if (!url) {
-                    toast.error('Failed to upload logo');
-                    return;
-                }
-                const ok = await updateLogo(url);
-                if (ok) onUpdated();
-            } catch (err) {
-                toast.error(
-                    `Failed to update logo: ${err instanceof Error ? err.message : 'Unknown error'}`
-                );
-            } finally {
-                setIsUploading(false);
-            }
-        };
-        upload();
-    }, [file, organization.id, uploadFile, updateLogo, onUpdated]);
 
     return (
         <div>
