+++ src/lib/components/settings/OrganizationTab.tsx
@@ -1,53 +1,35 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
 import TransferOwnershipModal from '@/lib/components/modal/TransferOwnershipModal';
 import DeleteOrganizationModal from '@/lib/components/modal/DeleteOrganizationModal';
 import { Button } from '@/lib/components/ui/Button';
 import { Input } from '@/lib/components/ui/Input';
-import { useAuth } from '@/lib/auth/auth-context';
-import { useOrgMembersAndInvites } from '@/lib/hooks/useOrgMembersAndInvites';
-import { useOrgSettings } from '@/lib/hooks/useOrgSettings';
 import OrgLogoUploader from '@/lib/components/settings/OrgLogoUploader';
+import useOrganizationTab from '@/lib/hooks/useOrganizationTab';
 
 export default function OrganizationTab() {
-    const router = useRouter();
-    const { activeOrganization, activeMember } = useAuth();
-    const orgId = activeOrganization?.id;
-    const isOwner = activeMember?.role === 'owner';
-
-    const { members } = useOrgMembersAndInvites(orgId);
-    const { renameOrg, updateLogo, isMutating, transferOwnership, deleteOrg } =
-        useOrgSettings(orgId);
-
-    const [nameDraft, setNameDraft] = useState(activeOrganization?.name ?? '');
-    const [transferOpen, setTransferOpen] = useState(false);
-    const [deleteOpen, setDeleteOpen] = useState(false);
-
-    useEffect(() => {
-        setNameDraft(activeOrganization?.name ?? '');
-    }, [activeOrganization?.name]);
+    const {
+        activeOrganization,
+        activeMember,
+        isOwner,
+        isMutating,
+        nameDraft,
+        setNameDraft,
+        renameDisabled,
+        handleRename,
+        updateLogo,
+        handleLogoUpdated,
+        transferOpen,
+        setTransferOpen,
+        deleteOpen,
+        setDeleteOpen,
+        eligibleTransferMembers,
+        transferOwnership,
+        deleteOrg,
+        redirectToDashboard,
+    } = useOrganizationTab();
 
     if (!activeOrganization) return null;
-
-    const trimmedName = nameDraft.trim();
-    const renameDisabled =
-        trimmedName === '' || trimmedName === activeOrganization.name || isMutating;
-
-    const handleRename = async () => {
-        if (renameDisabled) return;
-        const ok = await renameOrg(trimmedName);
-        if (ok) router.refresh();
-    };
-
-    const redirectToDashboard = () => {
-        router.push('/crm/dashboard');
-        router.refresh();
-    };
-
-    const eligibleTransferMembers =
-        isOwner && activeMember ? members.filter((m) => m.id !== activeMember.id) : [];
 
     return (
         <div className="flex flex-col gap-6">
@@ -62,7 +44,7 @@ export default function OrganizationTab() {
                     }}
                     updateLogo={updateLogo}
                     disabled={isMutating}
-                    onUpdated={() => router.refresh()}
+                    onUpdated={handleLogoUpdated}
                 />
 
                 <div className="col-start-2 flex min-w-0 flex-col gap-2">
